Removed ProjectInfo.getExtension()
This change starts a pattern of exposing some bits of the
DSL to ComponentImpl via VariantDslInfo. This restrict what
part of the DSL is visible and what isn't. For instance
Build Type and Product Flavors are not directly visible
from the DSL because they are filtered for the given variant
via properties on VariantDslInfo.

This change exposes some of these DSL objects to some
CreationConfig classes when there isn't a version of them
in the Variant API.

Bug: 202449978
Test: existing tests
Change-Id: Ia350195120ea3e8c67b0b198c443a72ff6c912ed