Handle executor being shut down
Previously, if we were about to start sending an update to studio and
right then a stop command was received (e.g. due to switching to
non-live mode), we could get a RejectedExecutionException which, since
it gets back to native code through the picture capture callback, would
cause JNI to kill the app.
Refactored so we'll probably not get the shut down executor at all, and
if we do the exception will be caught.
TODO is handling all exceptions that could be generated by this code, so
the app won't get killed for other reasons.

Test: Added a test
Fixes: 205595385
Change-Id: I0505c190be48e682ef14899c6f995f7e58cbe9fe