first iteration of SourceSetInputs cleanup.
SourceSetInputs is using Property<String> instead of just using
DirectoryProperty throughout. Replaced hard coded directory path
with usual InternalArtifactType.

The model files had to be changed because the path to the resValues
generated res is changed (with the introduction of the internal
artifact type) and is not optional and not present when the task
is not present (when the optional feature is turned off).

Test: Existing.
Bug: N/A
Change-Id: Ib93ba58543eda6da568b7a0f06abab24ec13ce8d