Apply implicit Kotlin attributes from properties
If you specify a Kotlin annotation without a usage site on a
constructor property or a class property, that annotation will *only*
be applied to the parameter, property or field, it will never be
applied to the generated getters and setters for the property. This
means that the annotations are not included in analysis of usages of
the property getters and setters.

This CL updates lint to (when possible) search for these additional
properties and include them in analysis; currently, this is true
for properties analyzed as source, but not for properties referenced
in bytecode.

Fixes: 200048369
Fixes: 148905488
Test: New unit tests included
Change-Id: Id5c0dc4a6b4ad9e75d2a35d67246a2de227add6a