Remove AOSP workarounds from lint
This CL removes old code which used to try to add some barebones
support for lint to work when running in an Android platform source
environment. However, this is no longer needed (lint is properly
integrated with the build system), and some of these workarounds are
directly harmful, such as the workaround for handling platform
annotations, which was breaking source offsets.

Test: Removes untested code
Bug: Not filed
Change-Id: I84676321351a2445148d2c9b5af7a782df721e16