Experimental flags to emulate final DSL flags for Ninja build system
This CL uses AGP's experimental flags to set up information equivalent to the what will be in the final DSL for this feature. We want to check in the actual DSL last because it will expose the Ninja build as a baked and documented feature.

After this CL, if the user tries to use the Ninja build system with these flags then they will receive a warning and the build will continue as if there is no Ninja build setup.

Bug: 192006965
Test: No behavior change that we want to enforce. Manually, confirmed warning that was added.
Change-Id: Ibdec5bf25c130888c0c08eee2ac0a47f2a3c16e8