Minimize the impact of profileable detector
The profileable detector attempts to perform method
sampling for every app process on Q & R to detect
whether it's profileable. If the process is new, the
method sampling can impact the app's startup.

This change tries to minimize the detector's impact:
(1) Make the sampling interval very long (30 minutes)
(2) Stop the method sampling as quickly as possible. It was
done by a child process before to avoid blocking the
profileable process detection, but the new app's startup
performance should have a higher priority.

Bug: 207159244
Test: manual
Change-Id: Ic77e9d4de3042ec36dc55897e18db0eeec944342