Add DefaultInfo provider, change compiletime file path
Some filegroup targets use java_import targets in their srcs, such as
tools/base/build-system/builder:agp_gradle_build_files. This is used to
build the gradle plugin, but does not work with jvm_import because we
did not include the jars in the DefaultInfo provider.

The compiletime file path needs to be updated as well, because
iml_to_build may create java_import rules which reference files outside
of the current package. Bazel does not allow rules to create files
outside of the target package, so we cannot use the 'sibling' parameter
of declare_file.

Bug: 200690965
Test: using jvm_import for all iml libraries
Change-Id: I1d9c6038c38434bb500ca308594fa2c409571813