Remove the @Beta annotations from the Lint APIs
The API is not guaranteed to be stable, and a large part of lint's API
surface is outside of our control (e.g. UAST, PSI, intellij-core), but
we have bytecode verification of custom lint jars now which pinpoints
incompatible changes, and we strive to not break APIs under our control,
so flagging all API accesses as unstable seems like overkill.

Test: N/A
Bug: None
Change-Id: I2b94ca37985f105ab78cd78b55a9d301f2c4e558